Add files via upload
I'm  going to read in, clean up, and visualize the real world project repository of Scala that spans data from a version control system (Git) as well as a project hosting site (GitHub). I'll find out who has had the most influence on its development and who are the experts.